[Curl][Soup] Move empty functions of ResourceHandle[Curl|Soup] and SynchronousLoaderClient[Curl|Soup] to ResourceHandle and SynchronousLoaderClient https://bugs.webkit.org/show_bug.cgi?id=258117
Reviewed by Michael Catanzaro.

The functions implemented in ResourceHandle[Curl|Soup].cpp and
SynchronousLoaderClient[Curl|Soup].cpp are not used in WebKit2,
so they are empty implementations and will not be added in the future.
Moving these functions reduces maintenance costs.  Also remove
unnecessary includes.

* Source/WebCore/loader/DocumentLoader.cpp:
* Source/WebCore/loader/FrameLoader.cpp:
* Source/WebCore/loader/LocalFrameLoaderClient.h:
* Source/WebCore/loader/PingLoader.cpp:
* Source/WebCore/loader/appcache/ApplicationCacheHost.cpp:
* Source/WebCore/loader/cache/CachedResource.cpp:
* Source/WebCore/page/Chrome.cpp:
* Source/WebCore/platform/Curl.cmake:
* Source/WebCore/platform/SourcesSoup.txt:
* Source/WebCore/platform/network/ResourceHandle.cpp:
(WebCore::ResourceHandle::~ResourceHandle):
(WebCore::ResourceHandle::start):
(WebCore::ResourceHandle::cancel):
(WebCore::ResourceHandle::platformSetDefersLoading):
(WebCore::ResourceHandle::platformLoadResourceSynchronously):
(WebCore::ResourceHandle::shouldUseCredentialStorage):
(WebCore::ResourceHandle::didReceiveAuthenticationChallenge):
(WebCore::ResourceHandle::receivedCredential):
(WebCore::ResourceHandle::receivedRequestToContinueWithoutCredential):
(WebCore::ResourceHandle::receivedCancellation):
(WebCore::ResourceHandle::receivedRequestToPerformDefaultHandling):
(WebCore::ResourceHandle::receivedChallengeRejection):
(WebCore::ResourceHandle::platformContinueSynchronousDidReceiveResponse): Deleted.
* Source/WebCore/platform/network/ResourceHandle.h:
* Source/WebCore/platform/network/SynchronousLoaderClient.cpp:
(WebCore::SynchronousLoaderClient::didReceiveAuthenticationChallenge):
(WebCore::SynchronousLoaderClient::platformBadResponseError):
* Source/WebCore/platform/network/curl/ResourceHandleCurl.cpp: Removed.
* Source/WebCore/platform/network/curl/SynchronousLoaderClientCurl.cpp: Removed.
* Source/WebCore/platform/network/soup/NetworkStorageSessionSoup.cpp:
* Source/WebCore/platform/network/soup/ResourceHandleSoup.cpp: Removed.
* Source/WebCore/platform/network/soup/SynchronousLoaderClientSoup.cpp: Removed.
* Source/WebKit/NetworkProcess/Downloads/Download.h:
* Source/WebKit/NetworkProcess/Downloads/DownloadManager.h:
* Source/WebKit/NetworkProcess/curl/NetworkDataTaskCurl.cpp:
* Source/WebKit/NetworkProcess/soup/NetworkProcessSoup.cpp:
* Source/WebKitLegacy/mac/WebCoreSupport/WebFrameLoaderClient.mm:

Canonical link: https://commits.webkit.org/265228@main
