+++ Source/WebCore/loader/LocalFrameLoaderClient.h
@@ -90,7 +90,6 @@ class ProtectionSpace;
 class RegistrableDomain;
 class RTCPeerConnectionHandler;
 class ResourceError;
-class ResourceHandle;
 class SecurityOrigin;
 class SharedBuffer;
 class SubstituteData;
