+++ Source/WebCore/page/Chrome.cpp
@@ -47,7 +47,6 @@
 #include "PageGroupLoadDeferrer.h"
 #include "PopupOpeningObserver.h"
 #include "RenderObject.h"
-#include "ResourceHandle.h"
 #include "Settings.h"
 #include "ShareData.h"
 #include "StorageNamespace.h"
