+++ Source/WebCore/platform/SourcesSoup.txt
@@ -29,11 +29,9 @@ platform/network/soup/CredentialSoup.cpp
 platform/network/soup/CredentialStorageSoup.cpp
 platform/network/soup/NetworkStorageSessionSoup.cpp
 platform/network/soup/ResourceErrorSoup.cpp
-platform/network/soup/ResourceHandleSoup.cpp
 platform/network/soup/ResourceRequestSoup.cpp
 platform/network/soup/ResourceResponseSoup.cpp
 platform/network/soup/SoupNetworkSession.cpp
-platform/network/soup/SynchronousLoaderClientSoup.cpp
 platform/network/soup/URLSoup.cpp
 platform/network/soup/WebKitFormDataInputStream.cpp @no-unify
 platform/network/soup/WebKitAutoconfigProxyResolver.cpp @no-unify
