+++ Source/WebCore/platform/network/ResourceHandle.cpp
@@ -175,13 +175,6 @@ void ResourceHandle::didReceiveResponse(ResourceResponse&& response, CompletionH
     client()->didReceiveResponseAsync(this, WTFMove(response), WTFMove(completionHandler));
 }
 
-#if !USE(SOUP) && !USE(CURL)
-void ResourceHandle::platformContinueSynchronousDidReceiveResponse()
-{
-    // Do nothing.
-}
-#endif
-
 ResourceRequest& ResourceHandle::firstRequest()
 {
     return d->m_firstRequest;
@@ -308,5 +301,71 @@ void ResourceHandle::setDefersLoading(bool defers)
 
     platformSetDefersLoading(defers);
 }
+
+#if USE(SOUP) || USE(CURL)
+ResourceHandleInternal::~ResourceHandleInternal() = default;
+
+ResourceHandle::~ResourceHandle()
+{
+    ASSERT_NOT_REACHED();
+}
+
+bool ResourceHandle::start()
+{
+    ASSERT_NOT_REACHED();
+    return false;
+}
+
+void ResourceHandle::cancel()
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::platformSetDefersLoading(bool)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::platformLoadResourceSynchronously(NetworkingContext*, const ResourceRequest&, StoredCredentialsPolicy, SecurityOrigin*, ResourceError&, ResourceResponse&, Vector<uint8_t>&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+bool ResourceHandle::shouldUseCredentialStorage()
+{
+    ASSERT_NOT_REACHED();
+    return false;
+}
+
+void ResourceHandle::didReceiveAuthenticationChallenge(const AuthenticationChallenge&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::receivedCredential(const AuthenticationChallenge&, const Credential&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::receivedRequestToContinueWithoutCredential(const AuthenticationChallenge&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::receivedCancellation(const AuthenticationChallenge&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::receivedRequestToPerformDefaultHandling(const AuthenticationChallenge&)
+{
+    ASSERT_NOT_REACHED();
+}
+
+void ResourceHandle::receivedChallengeRejection(const AuthenticationChallenge&)
+{
+    ASSERT_NOT_REACHED();
+}
+#endif
 
 } // namespace WebCore
