+++ Source/WebKit/NetworkProcess/Downloads/Download.h
@@ -50,7 +50,6 @@ class AuthenticationChallenge;
 class BlobDataFileReference;
 class Credential;
 class ResourceError;
-class ResourceHandle;
 class ResourceRequest;
 class ResourceResponse;
 }
