+++ Source/WebKit/NetworkProcess/Downloads/DownloadManager.h
@@ -43,7 +43,6 @@ class SessionID;
 
 namespace WebCore {
 class BlobDataFileReference;
-class ResourceHandle;
 class ResourceRequest;
 class ResourceResponse;
 }
